+++ cookbooks/acs-jenkins/metadata.rb
@@ -4,6 +4,6 @@ maintainer_email 'pdion891@apache.org'
 license          'Apache 2.0'
 description      'Installs/Configures acs-jenkins'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.0'
+version          '0.1.2'
 
 depends 'python'