+++ cookbooks/acs-jenkins/recipes/jdk18.rb
@@ -1,6 +1,6 @@
 #
 # Cookbook Name:: acs-jenkins
-# Recipe:: buildslave-centos
+# Recipe:: jdk18
 #
 # Copyright 2015, cloudstack.apache.org
 #
@@ -16,4 +16,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-package node['acs-jenkins']['pkgs']
+
+include_recipe 'java'
+
+#adding the below for the JDK8 installation
+java_ark "oracle-jdk8-x86_64" do
+  url node['java']['jdk']['8']['x86_64']['url']
+  app_home node['java']['jdk']['8']['install_path']
+  action :install
+end