+++ cookbooks/acs-jenkins/CHANGELOG.md
@@ -3,9 +3,17 @@ acs-jenkins CHANGELOG
 
 This file is used to list changes made in each version of the acs-jenkins cookbook.
 
+0.1.2
+-----
+- pdion891 - add expect as require package
+
+0.1.1
+-----
+- pdion891 - recipe jdk18
+
 0.1.0
 -----
-- [your_name] - Initial release of acs-jenkins
+- pdion891 - Initial release of acs-jenkins
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.